php: fix DESTDIR and other issues
php doesn't use DESTDIR but INSTALL_ROOT, which was misspelled.
this caused the installer to write into prefix directly.
since no manpath was passed to configure, it even created a directory /php
inside the prefix (thus it wrote it directly into /).
this is fixed as well.
there were also a huge amount of configure tests that produced a wrong
result due to ZEND failing to include the necessary headers for the
used functions (i.e. exit() without stdlib.h).