pango: hopefully the last fix for the pango module list generator
the program tried to look in /lib for the pango modules,
which are relocated there *after* the build.
so you got an invalid list on the first time building it,
and then everytime a valid one, since at this point it could
find the pango libs in /.

i added a quick hack into the program so you can pass a prefix=
argument. actually it should be called destdir, but it's too late
now.