gdb: build with -O0; cuts buildtime in half on 8 cores, it builds in 1:09 vs 2:13 before the change.
to remove the hardcoded -O2 -g which are *everywhere*
and can't be worked around; brute force is needed.

this makes gdb symbol lookup 15% slower.
since you're unlikely to debug HUGE C++ with millions of symbols,
this shouldn't matter at all.

the switch of optimization flags uncovered a symbol conflict in an
old patch that was needed before musl included dummies for dlopen
and friends for static linking.
thus the patch in question was removed.