prevent dummy package g++ from getting extracted
if a [build] section is specified, and mirrors available, the tarball
will be extracted. since it is not used, it can save a lot of time
to not do this unnecessary step.

(this package serves the single purpose to make the tarball getting
downloaded and its hash checked. the real work is done by the gcc4
package.)