libedit: fix segfaults and make links to libreadline.{so,a}
the el_set function (when used with some specific EL_ values)
expects a variable number of strings,
and a NULL as terminator to detect the end of the varargs.

at some places the NULL terminator was missing.
with this changes, libedit works as a gnu readline replacement,
at least for gdb (tested).