gdb: use libedit rather than the in-tree readline copy. this saves us some build time since libedit is going to be used later anyway. additionally libedit is 1/3 smaller than readline. however with 20 KLOC still far from optimal.
currently there's a minor issue with the built gdb:
it will not detect "DEL" keystrokes properly but instead insert
a tilde character, which is admittedly pretty uncomfortable.
however this could be the result of using termcap instead of
ncurses.
also when starting the libedit-linked gdb, there's a warning
about termcap db being not found.
this will be addressed soon.