rework stage0 build scripts so that stage0 and stage1 can use the same packages. build stage0 with the usual butch template.
to get it right, a new variable is "exported" from the butch
template: butch_root_dir.
it can be used to reference needed files (like musl-gcc) during
stage0 builds. the next step will be to merge stage0 and stage1
packages.

to get the butch template right, i dropped some support code
for the "stage U" idea, which was quite hacky. maybe i'll remove
it entirely or fix it to behave saner.