remove "relocate_toolchain"
this served the purpose to build a standalone toolchain in
/opt/toolchain. musl, gcc4, and binutils were built in there
with sysroot features, so that you could rip out the entire dir
and use it in a free-standing fashion on any compatible arch running
linux.
relocate_toolchain was the finalizer script that linked all files
from there into /.

as noble and useful as this was, it lead to bunch of ugly hacks
in the affected buildscripts. additionally the build template
had to be circumvented by quitting prematurely.

possibly the same result of a standalone toolchain can now be
achieved by simply copying the contents of
/opt/binutils
/opt/gcc4
/opt/musl
into a common root directory.
if not, it's now quite simple to use musl-cross to build something
very similar.