X11: fix the buggy behaviour
since this script is supposed to block, we can do not much more than
execing the Xfbdev with the right parameters.

with this configuration, you have to create an /etc/xinitrc
or ~/.xinitrc script with the following contents:

exec startlxde

to have it execute a proper window manager.
for more ideas about what to place in xinitrc see this here:
http://comptechdoc.org/os/linux/howlinuxworks/linux_hlxscripts.html