always build newest butch
i previously thought it's a good idea to save the time building it
but it turned out it's not.
it's only 1 minute at most even on slow hw, and i prefer having the
newest butch around. the skip-build only makes sense for the stage0
preparations, when it is possible that building butch would require
bootstrapping musl-gcc first to get posix_spawn on uclibc-based
hosts.