remove some cruft from config.cache
in order not to update the list constantly, it needs to behave more
whitelist-like.
it's safe to expose functions that are already implemented in musl,
but it's not safe to say it is not, because this could change soon.
i skipped cleaning up the sys/ headers for the moment, because i
think most of them are specific to some non-linux systems.