add qt4
i chose to use 4.3.4, the same version BLFS stable uses.
the reason is that this is the last release with a 40 MB tarball.
newer releases include webkit, and also the new documentation in
QtHelp's format, which is another 30 MB, for a total of 100-120MB.

this version is 2 years old, so i had to add some patches to be
compatible with recent libpng and openssl.