add xmms (1) (segfaults, at least on x86_64)
this is not the latest version, which is .11 as i found out after
making this package.

xmms1 uses glib1 and gtk1, so i convinced the build system that
those are available but made them use their 2.0 counterparts instead.
there's a total of 4 functions/variables that are no longer
available; i did some hacks to make them nops/try to patch them.

i have no idea if that broke it, but most of the patches should not
hurt. i have the feeling that the leader_window thing is serious
and should somehow be deactivated properly.

i'm adding this despite the segfault, since when started under gdb
once can actually see the XMMS window which is correctly drawn.
so there are hopes that only a minor fix is required to make it
work perfectly.
also there are a couple of warnings about invalid casts from pointers
to ints, so this issue may be 64bit-only.