add beep media player, v0.9.7
some devs forked xmms and converted it to gtk+2.
this is the result.

this version already looks very spartan.
but at least it compiles without pain, and should be compatible
with available plugins like uade.

there's some bug either in alsalib or in its alsa backend code
which causes a SIGFPE when opening mp3's.

i may later try to build an earlier version (096) as well;
because i hope that i get xmms1 + only the minimum fixes required
to get a working player.