gtk: remove obfuscated gtkalias symbols
this internal headers does some macro/hidden symbols tricks
that make regular functions start with IA__ prefix and
obfuscate what they're doing.
according to gtk+ devs, it's an "optimization".

however since it was removed from gtk+3, it seems it was a bogus
one ...

i'm removing it because i'm hunting a segfault in
gtk_file_chooser_dialog_new() and the symbol was magically renamed
in the debugger, causing a new layer of confusion and possible
bugs.