special-case gcc454 build to not cleanup DESTDIR
it's the default compiler, so links to gcc etc should keep working
until it is fully rebuilt.
otherwise trying to rebuild gcc454 would lead to no available
compiler on the system, as soon as the gcc454 package starts building
(except the stage0 compiler, but that one will not be available
before relocating into /).