alsa-utils: disable nls
for some reason, while building on arm, xgettext was detected but not
all the other gettext tools, and that lead to the all target being set
to all-yes in alsaconf/po (i.e. USE_NLS set to yes).
the lack of the msgfmt utility then caused some gmo file to not being
generated and make to fail.
i was unable to reproduce it on amd64.

disabling the nls support is needed anyway unless we want to
introduce a dependency on the gettext package.