sdl: set a reasonable default stack size
sdl's threads have no concept of stack size, so as a SDL user it is
completely outside your control.
thus we can't just rely on the implementation to give us a
sufficiently large stack, we have to request a reasonable
value for ourselves.

musl's default pthread stack size is 80KB, which is insufficient for
Dwarf Fortress, a rather large game relying on SDL_Threads.
96KB however is sufficient, so 128KB should be a good compatibility value
while not wasting huge amounts of mem.