bump butch to 0.3.0 and simplify build process
it's no longer needed to source the config before building stage0,
nor is it needed inside the rootfs before invoking butch.

butch is now being installed as "butch-core" and butch is a wrapper
script that sources the config.

that makes it easy to migrate the config into /etc/butch.conf at
some point.

an additional change of butch is that the package db, previously
/src/pkg/installed.dat, is now installed into /var/lib/butch.db
instead. that way /src/pkg/ can be removed without deleting the
package db by accident.
a reason why one may want to do that is to symlink pkg and KEEP
to a sabotage git checkout, so you automatically have fresh
packages after a git pull.

these changes require that $H is set to $PWD before the
stage0 config is sourced (automatically done by build-stage),
and that the config is aware of $H.
unfortunately this requires existing checkouts to update their config,
if they want to omit the ". config" step.