add "core" metapackage
it's planned to ship future sabotage images in a core and a full variant.
this is a first draft of what the core install will contain.
it should offer an environment that contains anything to get started
setting up network access and feature a complete development toolchain,
as well as a decent editor and a terminal-mode web browser, and maybe
some terminal irc client (but not irssi due to glib bloat).

so far, it's 440 MB uncompressed, which should be further optimized,
maybe by making curl use an alternate crypto library and other tweaks like
removing unneeded files, such as documentation in gnu info format.