add own su implementation
it turned out that busybox' su pulled in a lot of code from different
TUs, so it cannot easily be said if its attackable or not.
thus i implemented my own minimal su. it's been proof-read twice
by dalias, (and much more often by me) so it should be safe.
but feel free to study it as well and report any bug you see.

it's maintained separately in my "hardcore-utils" repo, but copied
into KEEP for simplicity.

solves half of #99
closes #32