kernel: add patch for '9pnet_virtio: no channels available' issue
when using
qemu-system-x86_64 -nographic -no-reboot -kernel $R/boot/vmlinuz \
-append "root=root ro rootflags=ro,trans=virtio,version=9p2000.L rootfstype=9p init=/bin/sh panic=1 HOST=x86_64 console=ttyS0" \
-fsdev local,id=root,path=$R,security_model=none \
-device virtio-9p-pci,fsdev=root,mount_tag=root

the above error message got displayed.
the cause is a mismatch between the devname passed to the p9_virtio_create
function and the names in the channel list in which devname is searched
in. the former comes prefixed with "/dev/", but the latter isn't.
the real bug is probably somewhere else, but this fixes the issue for now.