run-emulator: initialize kernel vga mode in non-console mode
now it's possible to run full X11 built from within chroot
using qemu.

the only fix needed is
sudo chmod 777 `utils/print-builddir`/share/X11/xkb/compiled/

for some reason, X11 wants write access to that file, and since
the rootfs was built inside a chroot with root privs, it's not
possible to write to it when qemu was started as regular user.

should help with #100