Revert "gcc*: also install libssp"
This reverts commit 812e7eced714e09fe88e44af7a3e52a19fceceb8.

It turned out that having libssp *installed* is not required,
since musl has everything needed already built-in.
despite that, at least one of the 2 changes in

1c8b161015b59e74a4486db4b9b15e1633cf4482

is needed to make gcc build the codegen support for the canary setup
code and the __stack_chk_fail call on function return.

the issue experienced while building openssh which led to the original
commit (now reverted) is described in #104 .