perl: debloat
the perl manpages are roughly 20MB.
usually the approach taken by sabotage is to install manpages,
however in this case it takes ca 30% of the entire install size.
additionally, the generated manpages use nasty syntax the manpager
can't deal very well with anyway.
and there's still perldoc, so only the documentation for built-in
modules is missing.
an interested perl developer using sabotage can still read them online,
or he can just edit the package file and undo the removal.