screen: put sockets into $HOME instead of /tmp
on suse, the socket is put into /var/run, which makes more sense
than /tmp.
imo it's even better to put them into home, so other users can not
even see other users sockets.
the only drawback of that is that screen will not work on a read-only
fs (where /var and /tmp are typically on tmpfs and thus writable).