thunderbird1: provide fake user agent, unless overridden by user
it's generally not a good idea to provide exact version of
client, host os and even kernel version - this makes a targetted
attack very easy - so we provide a fake one instead.

using a completely different user agent is not a good idea though,
as order and occurence of specific header tags can be used to
fingerprint an email client, so the fake user agent should at least
admit it's thunderbird, unless we carefully shuffle all other header
generation code around to match a certain mail client.

the user agent can still be overridden using the about:config options.