tzdata: install into /share/zoneinfo so musl can find it faster
musl's zoneinfo search order is /usr/share/zoneinfo, /share/zoneinfo,
and /etc/zoneinfo comes last. if someone uses a TZ env var that points
to a zoneinfo file, that means we need 2 syscalls more when localtime()
is first called.
since /usr is usually a symlink to /, we find it on the first strike
if it exists at all.