add shrunk down /etc/services
the perl package Net::Ping, used by CPAN, otherwise fails to
getservbyname("echo", "tcp") and raises an uncaught exception.

the full files /etc/services and /etc/protocols can be created using

http://sethwklein.net/iana-etc-2.30.tar.bz2
http://patches.cross-lfs.org/embedded-dev/iana-etc-2.30-update-2.patch

make get
make STRIP=yes

this only works when gawk is installed, thus we provide the service
directly with sabotage.
as the full services list is huge, which will cause musl to read
1000s of lines and thus slow down the system whenver getservbyname
or similar apis are used, we stripped most junk off.

this list here contains some protocols that are tcp only but
duplicated with a udp entry, those should be removed.

if someone needs a specific protocol in this list, we'll add it.
nmap seems to ship with its own list, so it works fine anyway.