change how services are installed
instead of copying everything from KEEP/etc/service, the services
now live in KEEP/services as single files, and are installed via
the services package. this makes it convenient to install new
services via butch update or butch rebuild services.

it's also nicer to have a single file for a service instead of a
directory tree, and the meta data (service down? does it need logging?)
inside the pkg description file.