default /dev/tty to 0666
that's how other distros and BSDs do it as well, and it apparently
is no security risk, as the kernel does the right thing for the
current terminal.

this fixes an issue with mksh as login shell for an ordinary user
when connecting via ssh as root and then su'ing to that user.