busybox: add patch(es) to support ping using SOCK_DGRAM
patch #1 updates ping to busybox as of 1.22.0
patch #2 fixes the incomplete SOCK_DGRAM (not upstream yet)

https://lkml.org/lkml/2011/5/13/382
http://openwall.info/wiki/people/segoon/ping
http://lists.busybox.net/pipermail/busybox/2014-January/080199.html
http://lists.busybox.net/pipermail/busybox/2014-January/080206.html

in order to activate the support, one has to modify
/proc/sys/net/ipv4/ping_group_range
(it takes 2 group ids, and any group within the range is allowed access.
for example 0 2000 -> all uids from0 to 2000 are allowed to use it)

this will be added to rc.local in the next commit.

closes #155
addresses #117