gcc3: fix libgcc visibility issues causing broken DSOs
gcc, when configured with --disable-shared, stupidly refuses
to mark the symbols from libgcc as hidden, which causes them to
creep into DSOs with exported visibility.

until recently gcc3 was only used to build static libs and apps,
however the merge of zlib and zlib-dynamic in fb8a9927a4a59b5ab
results in zlib's libz.so being built with gcc3 instead of gcc4,
very early during the stage1 build.
this caused failures (on i386) when the gcc3 built libz.so was
used during dropbear's configure tests (which ignore LDFLAGS)
and the build of openssl, and possibly others.

on x86_64, libgcc is basically unused so the issue went unnoticed.