butch template: make butch_do_relocate overridable by env
allows it to be set to false and prevent symlinking of a package,
so it's just built and stored in /opt/packagename.

this is handy if you want e.g. to build coreutils to use single
tools, but not overwrite the entire busybox install with it.

butch_do_relocate=false butch install coreutils