openssl: fix multiple issues inhibiting cross compilation
the most serious one was a makefile paralellism bug which led to
a try of linking the openssl binary with libcrypto.so that was
still in the process of getting written to (probably only when -flto
is active, which causes the library to take much longer to link).
doing a separate make invocation for the shared library stuff
ensures that the make process exits before launching another one
for the app targets.