add libstdc++ package
programs that are cross-compiled will end up with a hardcoded
reference to the cross-compiler's libstdc++ library,
and that breaks usage of --with-sysroot, which has the result
of adding the sysroot path to all library paths returned by libtool.
therefore we must explicitly add libstdc++ to the deps, to ensure
that the native gcc/libstdc++.so is built, which does not have this
problem.