add libtool.cross.gen
some packages ship a libtool (generated from ltmain.sh) which was generated
with an ancient libtoolize version, and are not compatible with cross-compiling
(they lack support for lt_sysroot).
some of these packages, most notably SDL 1.2.15 will likely not see any more
releases (end of life), so we can't hope that upstream gets it fixed.

adding such a huge script is quite contrary to sabotage's goals, however all
alternatives are considered worse.
those would be (among others)
- hack in buildscript to re-use libtool of some other package that uses a
  known-good libtool version (this would work as long as build dirs were not
  cleaned up recently, and no mix of native and x-compiled build dirs exist),
- make the offending packages depend on the entire auto* stack
  and regenerate the files (configure/libtool)
     $ libtoolize --automake
     $ aclocal -I ${SYSROOT}/usr/share/aclocal \
        [-I <dir_containing_your_project-specific_m4_macros>]
     $ autoconf
     $ autoheader
     $ automake -a
  however this depends on a ton a highly fragile/redundant/overdesigned stuff,
  which may even depend on gnu gettext (autopoint) and the existence of some
  3rd party m4 scripts in /share/aclocal...
- force the x-compile user to install a local "cross-libtool"
  ( http://metastatic.org/text/libtool.html ) which would be passed via LIBTOOL
  env var.
- adopt mlibtool and get it fixed for all usage cases of any package in sabotage
  so a fallback to the package's libtool never happens.

the third option seems the best of the bunch, however it puts yet another
complexity on the user's shoulders.
so we bite the bullet and add 260KB bloat to the repo...