introduce butch download template
this allows to set user-defined $wgetflags
(like retry, no certificate check, etc).
it also could be used to test for curl and use that in case it's available.
the template returns 44 in case a 404 is encountered.
this could be used to detect packages that have no single valid mirror url,
in which case retrying over and over to download could be caught in butch.

this first revision is deliberately kept simple.
as a first improvement, tarballs are downloaded to a temporary location first.
that way interrupted butch sessions (CTRL-C) will resume the tarball download
rather than erroring out due to size/hash mismatch.