fix apache "portable" runtime for x-compile
apr uses a braindead propietary pkg-config clone, which of course was
designed without a clue about crosscompiling or non-default install locations.
so trying to x-compile anything using apr is a PITA and requires a shitload
of hacks to fix.

they even ship their own libtool clone, which is of course buggy as well:
while it works for compiling, the link invocation calls a hardcoded "gcc"
rather than the linker name that was passed on the command line (would equal
to $CC).

as such, i had to use libtool.cross.gen to overwrite their libtool version,
however their usage case seems to hit a bashism in libtool.cross,
so we have to add bash as a dependency until libtool.cross gets fixed.