icu: prepare for x-compile
icu cannot be cross-compiled currently.
it wants to get a path set to an icu source dir that was compiled for the
host architecture.
so we'd have to build it twice for host and target.

configure: error: Error!
Cross compiling but no --with-cross-build option specified -
please supply the path to an executable ICU's build root