libtool.cross: fix for sdl-image
sdl-image uses a broken link command with $CC replaced with just "gcc".
since we added support for C++ to libtool, it behaves differently due to
the existance of "tags"; without them it swallows "gcc" and just silently
replaces it with the proper $CC. rather than hunting down the bug in
sdl-image's automake-generated mess, we fix libtool so it behaves as
expected by the author.