make perl packages fit for x-compile by disabling tests
"make test" assumes that the packages are built natively and can be
used with the perl binary in $PATH.
that's not true for cross-compilation when C sources are involved.
also the dependency chain of perl modules reachable via the host
perl interpreter usually does not match the ones in the cross-compile
sysroot.

whether the perl packages use the correct $CC as passed via the
environment is not clear yet, as their Makefiles default to non-verbose
mode.
additionally it's possible that some of them try to reference stuff
which is installed in the host directory structure and mess things up
that way.