kernel: bump to 3.14.23 and provide new kernel-tarball package
the intent is to be able to reference the kernel-tarball package from
separate standalone packages that make use of it, for example the
perf tool which is shipped as part of the kernel.

another novel mechanism is the usage of .version files in KEEP for
both the kernel and grsecurity. that way the version can be read
easily by auxiliary packages to reference the tarball.