gcc474: enable installation of libssp
this installs
libssp_nonshared.a
libssp.a
libssp.so*
into lib/.

the latter ones are probably unwanted, since musl already supplies the
regular libssp functionality. they will be removed in another commit.

libssp_nonshared.a is required for binaries compiled with
-fPIE -fstack-protector-all on some architectures, most notably i386,
where GOT lookup of functions requires multiple steps (instructions).
so, as an optimization, in order not to bloat the function prologue/epilogue
of every single stack-protected function unnecessarily,
it was decided to simply emit a jump to a local wrapper function which then
calls the real __stack_chk_fail function via GOT.
that's the reason the function has to be statically linked in and resides
in a separate library.
see ssp-local.c in gcc tree for the functions' code.

this is the first step in getting #274 solved.