gcc474: fix a line that returns error on rebuild (regression from enabling ssp)
gcc474 is one of the crucial packages blacklisted in the butch
build template, which must not be removed prior to rebuilding, since it
would render the system unfunctional.

since we enabled ssp in the previous commit, gcc installs a directory ssp
into its internal include dir in the DESTDIR.
this piece of code here copies the files from that include dir to another
location, in the assumption that they're only regular files.
as ssp is a directory, the cp command now returns error and displays the
message "omitting directory..." after a rebuild.

fixing it by just ignoring the error, as we dont need the ssp dir in the
copy target.