gcc474: use GCC_ARCH_CONFIG_FLAGS and autodetect if not set
the autodetection derives settings for arch, fpu model and float ABI.
this produces a toolchain that uses the same defaults the crosscompiler
uses, so the generated object files are always compatible and the
user isn't forced to specify the right flags on every gcc invocation.
this feature is important for the ARM architecture, which has a big
amount of different configurations.
now it should be possible to xcompile stage1 for armeabihf without issues.